requirements.txt
certifi==2018.4.16
PyYAML==3.12
SQLAlchemy==1.2.11
termcolor==1.1.0
